+++ projects/ngx-forms-inputs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @juulsgaard/ngx-forms-inputs
 
+## 17.1.1
+
+### Patch Changes
+
+- Update Input component typing
+- Updated dependencies
+  - @juulsgaard/ngx-forms@17.1.1
+
 ## 17.1.0
 
 ### Minor Changes